backEnd/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
six==1.16.0
tencentcloud-sdk-python==3.0.668
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
